requirements.txt
pypokerengine==1.0.1
numpy==1.24.0
scipy==1.10.0
scikit-learn==1.3.0
joblib==1.3.0
pyyaml==6.0
tqdm==4.65.0
matplotlib==3.7.1
cupy-cuda12x==13.5.1
fastrlock>=0.5

# Alternative GPU support:
# pip install cupy-cuda11x  # For CUDA 11.x (instead of cupy-cuda12x)
# OR install PyTorch with CUDA support:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118